src/cloud_functions/image_processor/requirements.txt
      
# Requirements for the Image Processor function
google-cloud-storage
google-cloud-aiplatform
earthengine-api
google-cloud-firestore   
